Changed IDs from BIGINT to INTEGER + fixed empty aws s3 error responses.

+++ src/classroom_has_courses/model/classroomhascourse.model.ts
@@ -20,7 +20,7 @@ import Grade from './../../grades/model/grade.model';
 export default class ClassroomHasCourse extends S.Model implements ClassroomHasCourseModel {
   @S.PrimaryKey
 	@S.AutoIncrement
-	@S.Column(S.DataType.BIGINT)
+	@S.Column(S.DataType.INTEGER)
 	public id!: number;
 
 	@S.AllowNull(true)
